+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 
 - fixed: incorrect `Could not determine base branch for PR` error messages when creating PRs from `git machete traverse -H`
 - fixed: `git bisect` is recognized as a separate repository state by `git machete status` and side-effecting operations
+- fixed: branches marked as `slide-out=no` are no longer slid out by `git machete slide-out --removed-from-remote`
 
 ## New in git-machete 3.36.0